[core_network] Bug fixes.

+++ core_network/src/commonMain/kotlin/com/varpihovsky/jetiqApi/data/Task.kt
@@ -25,5 +25,5 @@ class Task(
     val legend: String,
     @SerialName("num_mod")
     val module: String,
-    val points: Int,
+    val points: Float,
 )
